@@ -1,43 +1,52 @@
 {
-	"name": "laravel/laravel",
-	"description": "The Laravel Framework.",
-	"keywords": ["framework", "laravel"],
-	"license": "MIT",
-	"require": {
-		"laravel/framework": "4.1.*",
-                "unikent/curl": "dev-master",
-                "doctrine/dbal": "2.4.*"
-	},
-        "require-dev": {
-                 "way/generators": "2.*",
-                 "fzaninotto/faker": "v1.3.0",
-		 "barryvdh/laravel-ide-helper": "1.*"
-        },
-	"autoload": {
-		"classmap": [
-			"app/commands",
-			"app/controllers",
-			"app/models",
-			"app/database/migrations",
-			"app/database/seeds",
-			"app/tests/TestCase.php"
-		]
-	},
-	"scripts": {
-		"post-install-cmd": [
-			"php artisan clear-compiled",
-			"php artisan optimize"
-		],
-		"post-update-cmd": [
-			"php artisan clear-compiled",
-			"php artisan optimize"
-		],
-		"post-create-project-cmd": [
-			"php artisan key:generate"
-		]
-	},
-	"config": {
-		"preferred-install": "dist"
-	},
-	"minimum-stability": "stable"
+    "name": "laravel/laravel",
+    "description": "The Laravel Framework.",
+    "keywords": ["framework", "laravel"],
+    "license": "MIT",
+    "require": {
+        "laravel/framework": "4.2.*",
+        "unikent/curl": "dev-master",
+        "doctrine/dbal": "2.4.*",
+        "aws/aws-sdk-php": "~2.7",
+        "zizaco/entrust": "1.2.*@dev",
+        "monashee/php-simple-html-dom-parser": "dev-master"
+    },
+    "require-dev": {
+        "way/generators": "2.*",
+        "fzaninotto/faker": "v1.3.0",
+        "barryvdh/laravel-ide-helper": "1.*",
+        "codeception/codeception": "dev-master",
+        "panique/laravel-sass": "1.0"
+    },
+    "autoload": {
+        "classmap": [
+            "app/commands",
+            "app/controllers",
+            "app/models",
+            "app/database/migrations",
+            "app/database/seeds",
+            "app/tests/TestCase.php",
+            "app/pongo"
+        ],
+        "psr-4": {
+            "pongo\\": "app/pongo"
+        }
+    },
+    "scripts": {
+        "post-install-cmd": [
+            "php artisan clear-compiled",
+            "php artisan optimize"
+        ],
+        "post-update-cmd": [
+            "php artisan clear-compiled",
+            "php artisan optimize"
+        ],
+        "post-create-project-cmd": [
+            "php artisan key:generate"
+        ]
+    },
+    "config": {
+        "preferred-install": "dist"
+    },
+    "minimum-stability": "stable"
 }
