@@ -6,32 +6,32 @@ use Illuminate\Database\Schema\Blueprint;
 class CreateActionInstancesTable extends Migration
 {
 
-	/**
-	 * Run the migrations.
-	 *
-	 * @return void
-	 */
-	public function up()
-	{
-		Schema::create('action_instances', function(Blueprint $table) {
-			$table->increments('id');
-			$table->integer('action_id');
-			$table->integer('item_id');
-			$table->integer('session_id');
-			$table->timestamp('created');
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::create('action_instances', function(Blueprint $table) {
+            $table->increments('id');
+            $table->integer('action_id');
+            $table->integer('session_id');
+            $table->integer('item_id')->nullable();
+            $table->timestamp('created');
 
-			$table->foreign('action_id')->references('id')->on('actions')->onDelete('restrict')->onUpdate('cascade');
-		});
-	}
+            $table->foreign('action_id')->references('id')->on('actions')->onDelete('restrict')->onUpdate('cascade');
+        });
+    }
 
-	/**
-	 * Reverse the migrations.
-	 *
-	 * @return void
-	 */
-	public function down()
-	{
-		Schema::drop('action_instances');
-	}
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::drop('action_instances');
+    }
 
 }
