@@ -12,12 +12,12 @@
  */
 
 App::before(function($request) {
-	//
+//
 });
 
 
 App::after(function($request, $response) {
-	//
+//
 });
 
 /*
@@ -31,17 +31,111 @@ App::after(function($request, $response) {
   |
  */
 Route::filter('auth', function() {
-	if (Auth::guest())
-		return Redirect::guest('login');
+    if (Auth::guest())
+        return Redirect::guest('/');
+
+
+    $site_exists    = false;
+    $is_new_account = false;
+
+    if (Session::get("active_site_id") !== null) {
+        $site_exists = App\Models\Site::find(Session::get("active_site_id"))->count();
+        View::share(array("activeSiteName" => Session::get("active_site_name")));
+    }
+
+
+    $site_id   = $tenant_id = null;
+    if (Session::get("active_site_id") === null && !$site_exists) {
+        $site = App\Models\Site::where("account_id", Auth::user()->id)->get(array('id', 'name'))->first();
+        if ($site) {
+            $site_id   = $site->id;
+            $tenant_id = $site->name;
+
+            Session::set("active_site_id", $site->id);
+            Session::set("active_site_name", $site->name);
+            Session::remove("default_action_view");
+
+            $account_repository = new App\Pongo\Repository\AccountRepository();
+            $is_new_account     = $account_repository->isNewAccount();
+            Session::set('is_new_account', $is_new_account ? true : false);
+        }
+    }
+    else {
+        $site_id   = Session::get("active_site_id");
+        $tenant_id = Session::get("active_site_name");
+    }
+
+    if ($is_new_account) {
+        Session::remove('is_new_account');
+        return Redirect::to("v2/sites/{$tenant_id}/integration");
+    }
 });
 
 Route::filter('auth.basic', function() {
-	return Auth::basic();
+    return Auth::basic();
+});
+
+/**
+ * Filter of ajax request for data collection
+ */
+Route::filter('site.ajax', function($route) {
+
+    if (!\Request::ajax()) {
+        return Redirect::to('/');
+    }
+
+    $tenant_id = $route->parameter('tenant_id');
+
+    if (is_null($tenant_id)) {
+        return \Redirect::to('sites');
+    }
+
+    $validator = Validator::make(['name' => $tenant_id], ['name' => 'required|exists:sites,name']);
+    if ($validator->passes()) {
+        $repository = new \App\Pongo\Repository\SiteRepository();
+        $site       = $repository->isBelongToHim($tenant_id);
+        if (!$site)
+            return Redirect::to('/');
+
+        Session::set("active_site_id", $site->id);
+        Session::set("active_site_name", $site->name);
+    }
+    else
+        return Redirect::back()->withErrors($validator);
+});
+
+Route::filter('role.user', function() {
+    if (Auth::check()) {
+        if (!Auth::user()->hasRole('User'))
+            return Response::view('frontend.errors.missing', [], 404);
+    }
+    else
+        return Redirect::guest('/');
+});
+
+Route::filter('has.site', function () {
+    if (Auth::check()) {
+        $count_sites = App\Models\Account::find(Auth::user()->id)->sites()->count();
+        if ($count_sites <= 0) {
+            return Redirect::to('v2/sites/wizard');
+        }
+    }
+    else
+        return Redirect::guest('/');
+});
+
+Route::filter('role.admin', function() {
+    if (Auth::check()) {
+        if (!Auth::user()->hasRole('Administrator')) {
+            return Redirect::to('v2/home')->with("flash_error", "You are not allowed to access the page.");
+        }
+    }
+    else
+        return Redirect::guest('/');
 });
 
 App::missing(function($exception) {
-	return Response::view('frontend.errors.missing', array('exception' => $exception), 404);
-//	return Redirect::to('home');
+    return Response::view('frontend.errors.missing', array('exception' => $exception), 404);
 });
 
 
@@ -57,8 +151,8 @@ App::missing(function($exception) {
  */
 
 Route::filter('guest', function() {
-	if (Auth::check())
-		return Redirect::to('/');
+    if (Auth::check())
+        return Redirect::to('/');
 });
 
 /*
@@ -73,8 +167,7 @@ Route::filter('guest', function() {
  */
 
 Route::filter('csrf', function() {
-	if (Session::token() != Input::get('_token'))
-	{
-		throw new Illuminate\Session\TokenMismatchException;
-	}
+    if (Session::token() !== Input::get('_token')) {
+        throw new Illuminate\Session\TokenMismatchException;
+    }
 });
