@@ -7,23 +7,22 @@
  * Function     : 
  */
 return array(
-	"login"						 => "Login",
-	"login.with"				 => "Login with",
-	"or"						 => "Or",
-	"remember.me"				 => "Remember Me",
-	"forgot.password"			 => "Forgot Password",
-	"register"					 => "Register Account",
-	"reset.password"			 => "Reset Password",
-	"login"						 => "Login",
-	"login.now"					 => "Login now",
-	"info.pre.register_pref"	 => "By clicking register, you agree to the ",
-	"terms.and.conditions"		 => "Terms and Conditions",
-	"info.pre.register_suf"		 => " including our Cookie Use.",
-	"error.login.failed"		 => "Your email/password combination was incorrect.",
-	"error.email.doesnt.exists"	 => "Email you have entered doesn't exists.",
-	"signup.now"				 => "Signup Now",
-	"success.register"			 => "You have sucessfully registered. Please login.",
-	"success.password.changed"	 => "Password has been changed. You can login now.",
-	"subject.password.reminder"	 => "Password Reminder",
-	"info.top.register_pref"	 => "Already have an account. You can "
+    "login"                            => "Login",
+    "login.with"                       => "Login with",
+    "or"                               => "Or",
+    "remember.me"                      => "Remember Me",
+    "forgot.password"                  => "Forgot Password",
+    "register"                         => "Register Account",
+    "reset.password"                   => "Reset Password",
+    "login"                            => "Login",
+    "login.now"                        => "Login here",
+    "info.pre.register_pref"           => "By clicking register, you agree to the ",
+    "terms.and.conditions"             => "Terms and Conditions",
+    "info.pre.register_suf"            => " including our Cookie Use.",
+    "signup.now"                       => "Signup Now",
+    "success.register"                 => "You have sucessfully registered. Please login.",
+    "success.confirmation.correct"     => "Well done!. You've been confirmed to be part of us.",
+    "success.password.changed"         => "Password has been changed. You can login now.",
+    "subject.password.reminder"        => "Password Reminder",
+    "info.top.register_pref"           => "Already have an account ? You can "
 );
