@@ -7,37 +7,40 @@
  * Function     : 
  */
 return array(
-	"dashboard"				 => "Dashboard",
-	"create.site"			 => "Create Site",
-	"welcome"				 => "Welcome",
-	"settings"				 => "Settings",
-	"edit.profile"			 => "Edit Profile",
-	"edit.password"			 => "Edit Password",
-	"manage.sites"			 => "Manage Sites",
-	"manage.members"		 => "Manage Members",
-	"help"					 => "Help",
-	"logout"				 => "Logout",
-	"recommendations"		 => "Recommendations",
-	"create.recommendation"	 => "Create Recommendation",
-	"your.items"			 => "Your Items",
-	"overview"				 => "Overview",
-	"members"				 => "Members",
-	"items"					 => "Items",
-	"statistics"			 => "Statistics",
-	"actions"				 => "Actions",
-	"rulesets"				 => "Rulesets",
-	"filters"				 => "Filters",
-	"widgets"				 => "Widgets",
-	"overview"				 => "Overview",
-	"add.new"				 => "Add New",
-	"manage"				 => "Manage",
-	"yes"					 => "Yes",
-	"no"					 => "No",
-	"edit"					 => "Edit",
-	"delete"				 => "Delete",
-	"remove"				 => "Remove",
-	"add"					 => "Add",
-	"view"					 => "View",
-	"remove.confirm"		 => "Are you sure want to remove this?",
-	""						 => "",
+    "dashboard"             => "Dashboard",
+    "create.site"           => "Create Site",
+    "welcome"               => "Welcome",
+    "settings"              => "Settings",
+    "edit.profile"          => "Edit Profile",
+    "edit.password"         => "Edit Password",
+    "manage.sites"          => "Manage Sites",
+    "manage.members"        => "Manage Members",
+    "help"                  => "Help",
+    "logout"                => "Logout",
+    "recommendations"       => "Recommendations",
+    "create.recommendation" => "Create Recommendation",
+    "your.items"            => "Your Items",
+    "overview"              => "Overview",
+    "members"               => "Members",
+    "items"                 => "Items",
+    "statistics"            => "Statistics",
+    "actions"               => "Actions",
+    "rulesets"              => "Rulesets",
+    "filters"               => "Filters",
+    "widgets"               => "Widgets",
+    "overview"              => "Overview",
+    "add.new"               => "Add New",
+    "manage"                => "Manage",
+    "yes"                   => "Yes",
+    "no"                    => "No",
+    "edit"                  => "Edit",
+    "delete"                => "Delete",
+    "remove"                => "Remove",
+    "add"                   => "Add",
+    "view"                  => "View",
+    "remove.confirm"        => "Are you sure want to remove this?",
+    "edit.business"         => "Edit Business",
+    "site.updated"          => "Site business has been updated.",
+    "add.new.site"          => "Add New Site",
+    ""                      => "",
 );
