@@ -1,19 +1,19 @@
 <?php
 
 return array(
-	/*
-	  |--------------------------------------------------------------------------
-	  | Password Reminder Language Lines
-	  |--------------------------------------------------------------------------
-	  |
-	  | The following language lines are the default lines which match reasons
-	  | that are given by the password broker for a password update attempt
-	  | has failed, such as for an invalid token or invalid new password.
-	  |
-	 */
+    /*
+      |--------------------------------------------------------------------------
+      | Password Reminder Language Lines
+      |--------------------------------------------------------------------------
+      |
+      | The following language lines are the default lines which match reasons
+      | that are given by the password broker for a password update attempt
+      | has failed, such as for an invalid token or invalid new password.
+      |
+     */
 
-	"password"	 => "Passwords must be at least six characters and match the confirmation.",
-	"user"		 => "We can't find a user with that e-mail address.",
-	"token"		 => "This password reset token is invalid.",
-	"sent"		 => "Password reminder sent!",
+    "password" => "Passwords must be at least six characters and match the confirmation.",
+    "user"     => "We can't find a user with that e-mail address.",
+    "token"    => "This password reset token is invalid.",
+    "sent"     => "Password reminder sent!",
 );
