@@ -10,21 +10,24 @@
 
 namespace App\Models;
 
-class ActionMeta extends \Eloquent
+use Eloquent;
+
+class ActionMeta extends Eloquent
 {
 
-	/**
-	 * The database table used by the model.
-	 *
-	 * @var string
-	 */
-	public $timestamps	 = false;
-	protected $table	 = 'action_metas';
+    /**
+     * The database table used by the model.
+     *
+     * @var string
+     */
+    protected $guarded = array("id");
+    public $timestamps = false;
+    protected $table   = 'action_metas';
 
-	public function action()
-	{
-		return $this->hasMany("App\Models\Action", "action_id");
-	}
+    public function action()
+    {
+        return $this->hasMany("App\Models\Action", "action_id");
+    }
 
 }
 
