@@ -10,15 +10,21 @@
 
 namespace App\Models;
 
+use Illuminate\Database\Eloquent\SoftDeletingTrait;
+
 class Combination extends \Eloquent
 {
 
-	/**
-	 * The database table used by the model.
-	 *
-	 * @var string
-	 */
-	protected $table = 'combinations';
+    use SoftDeletingTrait;
+
+    protected $dates = ['deleted_at'];
+
+    /**
+     * The database table used by the model.
+     *
+     * @var string
+     */
+    protected $table = 'combinations';
 
 }
 
