@@ -13,16 +13,28 @@ namespace App\Models;
 class Filter extends \Eloquent
 {
 
-	/**
-	 * The database table used by the model.
-	 *
-	 * @var string
-	 */
-	protected $table			 = 'filters';
-	public $manage_table_header	 = array(
-		"name"		 => "Filter Name",
-		"properties" => "Selected Properties"
-	);
+    /**
+     * The database table used by the model.
+     *
+     * @var string
+     */
+    protected $table            = 'filters';
+    public $manage_table_header = array(
+        "name"       => "Filter Name",
+        "properties" => "Selected Properties"
+    );
+    public $filter_data_type    = [
+        'str'      => 'text',
+        'num'      => 'numeric',
+        'date'     => 'date',
+        'list'     => 'list',
+        'location' => 'location'
+    ];
+
+    public function metas()
+    {
+        return $this->hasMany("App\Models\FilterMeta");
+    }
 
 }
 
