@@ -10,15 +10,15 @@
 
 namespace App\Models;
 
-class Filtermeta extends \Eloquent
+class FilterMeta extends \Eloquent
 {
 
-	/**
-	 * The database table used by the model.
-	 *
-	 * @var string
-	 */
-	protected $table = 'filter_metas';
+    /**
+     * The database table used by the model.
+     *
+     * @var string
+     */
+    protected $table = 'filter_metas';
 
 }
 
