@@ -13,29 +13,29 @@ namespace App\Models;
 class Member extends \Eloquent
 {
 
-	/**
-	 * The database table used by the model.
-	 *
-	 * @var string
-	 */
-	public $timestamps			 = false;
-	protected $primaryKey		 = "id";
-	protected $table			 = 'members';
-	public $manage_table_header	 = array(
-		"name"		 => "Name",
-		"email"		 => "Email",
-		"created_at" => "Date Created"
-	);
-	public $rules				 = array(
-		"name"		 => "required",
-		"email"		 => "required|email|unique:accounts|unique:members",
-		"password"	 => "required|min:8|confirmed"
-	);
+    /**
+     * The database table used by the model.
+     *
+     * @var string
+     */
+    public $timestamps          = false;
+    protected $primaryKey       = "id";
+    protected $table            = 'members';
+    public $manage_table_header = array(
+        "name"       => "Name",
+        "email"      => "Email",
+        "created_at" => "Date Created"
+    );
+    public $rules               = array(
+        "name"     => "required",
+        "email"    => "required|email|unique:accounts|unique:members",
+        "password" => "required|min:8|confirmed"
+    );
 
-	public function detail()
-	{
-		return $this->belongsTo("App\Models\Account", "account_id");
-	}
+    public function detail()
+    {
+        return $this->belongsTo("App\Models\Account", "account_id");
+    }
 
 }
 
