@@ -13,25 +13,21 @@ namespace App\Models;
 class Session extends \Eloquent
 {
 
-	/**
-	 * The database table used by the model.
-	 *
-	 * @var string
-	 */
-	protected $table = 'sessions';
+    protected $guarded = array("id");
+    protected $table   = 'sessions';
 
-	public function visitor()
-	{
-		return $this->belongsTo("App\Models\Visitor");
-	}
+    public function visitor()
+    {
+        return $this->belongsTo("App\Models\Visitor");
+    }
 
-	static function getNumberOfUsers($site_id)
-	{
-		$query			 = "SELECT COUNT(*) FROM "
-				. "(SELECT count(distinct visitor_id) FROM sessions WHERE site_id = {$site_id} GROUP BY visitor_id) AS tbl";
-		$total_actions	 = \DB::select($query);
-		return $total_actions[0]->count;
-	}
+    static function getNumberOfUsers($site_id)
+    {
+        $query         = "SELECT COUNT(*) FROM "
+                . "(SELECT count(distinct visitor_id) FROM sessions WHERE site_id = {$site_id} GROUP BY visitor_id) AS tbl";
+        $total_actions = \DB::select($query);
+        return $total_actions[0]->count;
+    }
 
 }
 
