@@ -1,13 +1,17 @@
 <!DOCTYPE html>
 <html lang="en-US">
-	<head>
-		<meta charset="utf-8">
-	</head>
-	<body>
-		<h2>Welcome to Predictry, {{ $fullname }}!</h2>
+    <head>
+        <meta charset="utf-8">
+    </head>
+    <body>
+        <h2>Verify Your Email Address</h2>
 
-		<div>
-			Your account has been successfully created. <a href="{{ URL::to('login') }}">Login</a> to access your account now.
-		</div>
-	</body>
+        <div>
+            Thanks for creating an account on Predictry Website.
+            Please follow the link below to verify your email address
+            {{ URL::to('v2/verify/' . $confirmation_code) }}.<br/>
+
+        </div>
+
+    </body>
 </html>
