@@ -1,13 +1,13 @@
 <!DOCTYPE html>
 <html lang="en-US">
-	<head>
-		<meta charset="utf-8">
-	</head>
-	<body>
-		<h2>Password Reset</h2>
+    <head>
+        <meta charset="utf-8">
+    </head>
+    <body>
+        <h2>Password Reset</h2>
 
-		<div>
-			To reset your password, complete this form: {{ URL::to('password/reset', array($token)) }}.
-		</div>
-	</body>
+        <div>
+            To reset your password, complete this form: {{ URL::to('v2/password/reset', array($token)) }}.
+        </div>
+    </body>
 </html>
