@@ -1,5 +1,6 @@
 </div>
 <p class="text-center">&copy; 2014 Predictry. A <a href='http://www.vventures.asia'>V Ventures</a> company. <br/> <span class="small cl-fade"> Made with <i class="fa fa-heart-o"></i> in KL</span></p>
+<p class="text-center"><small class="text-muted">Version 0.2.5</small></p>
 </div>
 {{ HTML::script('assets/js/jquery-1.11.0.min.js') }}
 {{ HTML::script('assets/js/bootstrap.min.js') }}
@@ -9,7 +10,6 @@
 {{ $script or 'default' }}
 @endforeach
 @endif
-{{ HTML::script('assets/js/script.js') }}
 {{ $custom_script or 'default' }}
 </body>
 </html>
