@@ -1,13 +1,13 @@
 @if(Session::has('flash_error'))
 <div class="alert alert-danger alert-dismissable">
-	<button type="button" class="close" data-dismiss="alert" aria-hidden="true">&times;</button>
-	<strong>Error!</strong> {{ Session::get('flash_error') }}
+    <button type="button" class="close" data-dismiss="alert" aria-hidden="true">&times;</button>
+    {{ Session::get('flash_error') }}
 </div>
 @endif
 
 @if(Session::has('flash_message'))
 <div class="alert alert-success">
-	<button type="button" class="close" data-dismiss="alert" aria-hidden="true">&times;</button>
-	<strong>Congratulation!</strong> {{ Session::get('flash_message') }}
+    <button type="button" class="close" data-dismiss="alert" aria-hidden="true">&times;</button>
+    {{ Session::get('flash_message') }}
 </div>
 @endif
