@@ -1,12 +1,13 @@
 <div class="modal fade" id="viewModal">
-	<div class="modal-dialog">
-		<div class="modal-content">
-			<div class="modal-header">
-				<button type="button" class="close" data-dismiss="modal" aria-hidden="true">&times;</button>
-				<h4 class="modal-title">{{ $modalTitle or '&nbsp;' }}</h4>
-			</div>
-			<div class="modal-body">
-			</div>
-		</div><!-- /.modal-content -->
-	</div><!-- /.modal-dialog -->
+    <div class="modal-dialog">
+        <div class="modal-content">
+            <div class="modal-header">
+                <button type="button" class="close" data-dismiss="modal" aria-hidden="true">&times;</button>
+                <h4 class="modal-title">{{ $modalTitle or '&nbsp;' }}</h4>
+            </div>
+            <div class="modal-body">
+                {{ $modalContent or '' }}
+            </div>
+        </div><!-- /.modal-content -->
+    </div><!-- /.modal-dialog -->
 </div><!-- /.modal -->
